// src/app/components/models/all.models.ts
export interface Student {
  id: number;
  name: string;
  surName: string;
  rollNo: number;
  maktabClass?: MaktabClass | null;
  age: number;
  gender: 'Male' | 'Female';
  guardianName?: string;
  phone?: string;
  imagePath?: string;
  fees: Fee[];
  status: string;
  admissionDate: string;
  attendance: Attendance[];
  dob: string;
  fatherName?: string;
  surname?: string;
  whatsapp?: string;
  classId?: number;
  address?: string;
  schoolName?: string;
  parentProfession?: string;
  preferredTime?: string;
}

// src/app/models/maktab-class.model.ts
export interface MaktabClass {
  studentCount: number;
  timing: string;
  division: string;
  id?: number;
  name: string;
  teacher?: Teacher[];
  subject?: Subject[];
  student?: Student[];
}

// src/app/models/teacher.model.ts
export interface Teacher {
  maktabClass: MaktabClass;
  id?: number;
  fullName: string;
  phone?: string;
}

// src/app/models/subject.model.ts
export interface Subject {
  id?: number;
  name: string;
  classes?: MaktabClass[];
  teacher?: Teacher[];
}

// src/app/models/attendance.model.ts
export interface Attendance {
  id: number;
  studentId: number;
  studentName: string;
  student: Student;
  className: string;
  attendanceDate: string; // YYYY-MM-DD
  attendanceStatus: 'PRESENT' | 'ABSENT' | 'LATE';
  remark: string;
  minutesLate: number | null;
}

// src/app/models/fee.model.ts
export interface Fee {
  id: number;
  studentId: number;
  month: string;
  year: number;
  status: string;
  paymentDate: Date;
  dueDate: Date;
  paymentMode: string;
  recordedBy: string;
  remark: string;
  paidAmount: number;
}